Update index.js
The Next Link component is used to enable client-side transitions between routes of the same app but in this case we are linking to an external page, therefore an <a> is better suited. 
The target blank attribute to open a new page is also a good functionality because you may want to keep this page open and open the GameStop NFT url in another window.